+++ apps/earn/components/PoolsSection/Tables/PoolRewardsCell.tsx
@@ -3,7 +3,7 @@ import { SUSHI } from '@sushiswap/currency'
 import { Currency } from '@sushiswap/ui'
 import { FC } from 'react'
 
-import { ICON_SIZE } from './contants'
+import { ICON_SIZE } from './constants'
 import { Row } from './SharedCells/types'
 
 export const PoolRewardsCell: FC<Row<{ chainId: ChainId }>> = ({ row }) => {