+++ apps/earn/components/RemoveSection/RemoveSectionUnstake.tsx
@@ -2,7 +2,7 @@ import { tryParseAmount } from '@sushiswap/currency'
 import { ChefType, Pool, usePool } from '@sushiswap/client'
 import { useIsMounted } from '@sushiswap/hooks'
 import { AppearOnMount, Dots } from '@sushiswap/ui'
-import { getMasterChefContractConfig, useMasterChefWithdraw } from '@sushiswap/wagmi'
+import { useMasterChefWithdraw } from '@sushiswap/wagmi'
 import { FC, useMemo, useState } from 'react'
 
 import { useGraphPool } from '../../lib/hooks'
@@ -46,7 +46,6 @@ export const _RemoveSectionUnstake: FC<AddSectionStakeProps> = withCheckerRoot((
     data: { reserve0, reserve1, liquidityToken },
   } = useGraphPool(pool)
   const { balance } = usePoolPositionStaked()
-  const { approved } = useApproved(APPROVE_TAG_UNSTAKE)
   const amount = useMemo(() => {
     return tryParseAmount(value, liquidityToken)
   }, [liquidityToken, value])
@@ -56,7 +55,6 @@ export const _RemoveSectionUnstake: FC<AddSectionStakeProps> = withCheckerRoot((
     amount,
     pid: farmId,
     chef: chefType,
-    enabled: approved,
   })
 
   return (
@@ -78,27 +76,16 @@ export const _RemoveSectionUnstake: FC<AddSectionStakeProps> = withCheckerRoot((
               </Button>
             }
           >
-            <Checker.ApproveERC20
-              size="xl"
+            <Button
+              onClick={() => sendTransaction?.()}
               fullWidth
-              id="unstake-approve-slp"
-              amount={amount}
-              contract={getMasterChefContractConfig(pool.chainId, chefType)?.address}
-              enabled={Boolean(getMasterChefContractConfig(pool.chainId, chefType)?.address)}
+              size="xl"
+              variant="filled"
+              disabled={isWritePending}
+              testId="unstake-liquidity"
             >
-              <Checker.Success tag={APPROVE_TAG_UNSTAKE}>
-                <Button
-                  onClick={() => sendTransaction?.()}
-                  fullWidth
-                  size="xl"
-                  variant="filled"
-                  disabled={!approved || isWritePending}
-                  testId="unstake-liquidity"
-                >
-                  {isWritePending ? <Dots>Confirm transaction</Dots> : 'Unstake Liquidity'}
-                </Button>
-              </Checker.Success>
-            </Checker.ApproveERC20>
+              {isWritePending ? <Dots>Confirm transaction</Dots> : 'Unstake Liquidity'}
+            </Button>
           </Checker.Custom>
         </Checker.Network>
       </Checker.Connect>