+++ packages/wagmi/src/systems/PoolFinder/types.ts
@@ -1,7 +1,7 @@
 import { ConstantProductPool, Fee, Pair, StablePool } from '@sushiswap/amm'
 import { BentoBoxV1ChainId } from '@sushiswap/bentobox'
 import { Type } from '@sushiswap/currency'
-import { UniswapV2Router02ChainId } from '@sushiswap/v2-core'
+import { SushiSwapV2ChainId } from '@sushiswap/v2-sdk'
 import { ReactElement } from 'react'
 
 import { ConstantProductPoolState, PairState, StablePoolState } from '../../hooks'
@@ -23,7 +23,7 @@ interface PoolFinderProps {
 }
 
 export interface LegacyPoolFinderProps extends PoolFinderProps {
-  chainId: UniswapV2Router02ChainId
+  chainId: SushiSwapV2ChainId
   token0: Type | undefined
   token1: Type | undefined
   index?: number