+++ config/react-query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sushiswap/wagmi-config
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`c6d6cc5db`](https://github.com/sushiswap/sushiswap/commit/c6d6cc5db4cc614f3931ee3a325547967c86c51a)]:
+  - @sushiswap/chain@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes